+++ src/presenters/entity-page-pinned-projects.jsx
@@ -11,68 +11,9 @@ import AddProjectToCollection from './includes/add-project-to-collection.jsx';
 
 /* globals Set */
 
-function trackRemix(id, domain) {
-  analytics.track("Click Remix", {
-    origin: "project page",
-    baseProjectId: id,
-    baseDomain: domain,
-  });
-}
-
-const FeaturedProject = ({api, isAuthorized, currentUser, unfeatureProject, addProjectToCollection, featuredProject}) => {
+const EntityPagePinnedProjects = ({api, projects, currentUser, isAuthorized, removePin, projectOptions, addProjectToCollection,}) => {
   
-  const reportBtn = 
-    <div className="buttons buttons-left">
-      <ReportButton className="button-small" reportedType="project" reportedModel={featuredProject} />
-    </div>;
-  
-  return(
-    <>
-      <section id="featured-project-embed">            
-        {isAuthorized && <FeaturedProjectOptionsPop api={api} unfeatureProject={unfeatureProject} featuredProjectId={featuredProject.id}/>}
-        <div className="glitch-embed-wrap">
-          <iframe title="embed"
-            src={`${APP_URL}/embed/#!/embed/${featuredProject.id}?path=README.md&previewSize=100`}
-            allow="geolocation; microphone; camera; midi; encrypted-media"
-          ></iframe>
-        </div>
-
-        {isAuthorized ?
-          <div className="buttons buttons-left">
-            <EditButton className="button-small button-edit" name={featuredProject.id} isMember={isAuthorized}/>
-          </div>
-          : reportBtn
-        }
-
-        <div className="buttons buttons-right">
-
-          {currentUser.login && <AddProjectToCollection className="button-small" api={api} currentUser={currentUser} project={featuredProject} fromProject={true} addProjectToCollection={addProjectToCollection}/>}
-          
-          <RemixButton className="button-small"
-            name={featuredProject.domain} isMember={isAuthorized}
-            onClick={() => trackRemix(featuredProject.id, featuredProject.domain)}
-          />
-        </div>
-      </section>
-    </>
-  );
-};
-
-FeaturedProject.propTypes = {
-  api: PropTypes.func,
-  isAuthorized: PropTypes.bool.isRequired,
-  currentUser: PropTypes.object.isRequired,
-  unfeatureProject: PropTypes.func,
-  featuredProject: PropTypes.object,
-  addProjectToCollection: PropTypes.func,
-};
-
-const EntityPagePinnedProjects = ({api, projects, pins, currentUser, isAuthorized, removePin, projectOptions, featuredProjectId, addProjectToCollection,}) => {
-  const pinnedSet = new Set(pins.map(({projectId}) => projectId));
-  const pinnedProjects = projects.filter( ({id}) => pinnedSet.has(id)).filter ( ({id}) => id != featuredProjectId); 
-  const featuredProject = Object.is(featuredProjectId, undefined) ? undefined : projects.filter( ({id}) => id == featuredProjectId)[0];
-  
-  const pinnedVisible = (isAuthorized || pinnedProjects.length) && projects.length;
+  const pinnedVisible = (isAuthorized || projects.length);
     
   const pinnedTitle = (
     <>
@@ -82,32 +23,17 @@ const EntityPagePinnedProjects = ({api, projects, pins, currentUser, isAuthorize
   );
   
   return (
-    <>
-      {!!pinnedVisible && (!!pinnedProjects.length || featuredProjectId) && (
-       <>       
-       <h2 className="pinned">{pinnedTitle}</h2>
+     {projects.length > 0 && 
+        <ProjectsList title={pinnedTitle}
+          projects={pinnedProjects}
+          api={api} 
+          projectOptions={isAuthorized ? {removePin, ...projectOptions} 
+            : (currentUser && currentUser.login ? {...projectOptions} : {})
+          }
+          extraClasses="pinned"
+        />
+     }
        
-         {featuredProjectId && 
-            <FeaturedProject   
-              {...{api, isAuthorized, currentUser, addProjectToCollection}}
-              unfeatureProject={projectOptions.unfeatureProject}
-              featuredProject={featuredProject}
-            />
-         }
-
-         {pinnedProjects.length > 0 && 
-            <ProjectsList title=""
-              projects={pinnedProjects}
-              api={api} 
-              projectOptions={isAuthorized ? {removePin, ...projectOptions} 
-                : (currentUser && currentUser.login ? {...projectOptions} : {})
-              }
-              extraClasses="pinned"
-            />
-         }
-       </>
-      )}
-    </>
   );
 };
 EntityPagePinnedProjects.propTypes = {
@@ -116,9 +42,6 @@ EntityPagePinnedProjects.propTypes = {
   currentUser: PropTypes.object,
   isAuthorized: PropTypes.bool.isRequired,
   projects: PropTypes.array.isRequired,
-  pins: PropTypes.arrayOf(PropTypes.shape({
-    projectId: PropTypes.string.isRequired,
-  }).isRequired).isRequired,
   removePin: PropTypes.func.isRequired,
   projectOptions: PropTypes.object,
 };
